requirements.txt
numpy==1.20.1
tensorflow-gpu==1.15
pillow==8.1.2
requests==2.25.1
selenium==3.141.0
pycocotools==2.0.2
six==1.15.0
opencv-python==4.5.1.48
setuptools==52.0.0
h5py==3.2.1
flask==1.1.2
waitress==2.0.0